requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
pytest
pytest-asyncio
pytest-cov
pytest-mock
httpx
factory-boy
faker